+++ geoparser/annotator/templates/annotate.html
