+++ geoparser/annotator/templates/continue_session.html
