+++ geoparser/cli/annotator.py
@@ -1,4 +1,4 @@
-from geoparser.app import run
+from geoparser.annotator import run
 
 
 def annotator_cli():
